// sp/ClientApp/src/app/models/contacts/contact-model.ts
export class ContactModel {
    friendName: string;
    firstName: string;
    location: string;
    picturePath: string;
    connectionID: string;
    showType: string;
    status: string;
    titleDesc: string;
    Params: string;
    paramsAV: string;
    email: string;
    labelText: string;
    nameAndID: string;
}